Remove images from local cache on new import, and only import oisp images.

+++ util/restart-cluster.sh
@@ -3,8 +3,12 @@
 SCRIPTDIR=$(dirname "$0")
 CLUSTERNAME=${CLUSTERNAME:-oispcluster}
 
+if [ -z "$K3S_IMAGE" ]
+then
+	>&2 echo "ERROR: Image version must be specified in K3S_IMAGE env variable or the script must be called from the Makefile" && exit 1;
+fi
+
 k3d cluster delete ${CLUSTERNAME} || echo "Cluster ${CLUSTERNAME} could not be deleted/ does not exist"
-# Image version must be specified as an environment variable or the script must be called from the Makefile
 k3d cluster create ${CLUSTERNAME} --registry-use k3d-oisp.localhost:12345 --image=${K3S_IMAGE}
 
 printf "\033[1mKubernetes cluster started\033[0m\n"
